+++ src/errors.rs
@@ -28,7 +28,7 @@ pub enum ClientError {
     StoreError(StoreError),
     TransactionExecutorError(TransactionExecutorError),
     TransactionProvingError(TransactionProverError),
-    ExistanceVerificationError(NoteId),
+    ExistenceVerificationError(NoteId),
 }
 
 impl fmt::Display for ClientError {
@@ -64,7 +64,7 @@ impl fmt::Display for ClientError {
             ClientError::TransactionProvingError(err) => {
                 write!(f, "transaction prover error: {err}")
             },
-            ClientError::ExistanceVerificationError(note_id) => {
+            ClientError::ExistenceVerificationError(note_id) => {
                 write!(f, "The note with ID {note_id} doesn't exist in the chain")
             },
         }